requirements.txt
langchain==0.3.1
openai==1.51.0
python-dotenv==1.0.1
langchain-openai==0.2.1